requirements.txt
flask
requests
beautifulsoup4
lxml
python-dotenv
flask-cors
gunicorn
pyyaml
transformers
llama_index
pinecone-client